% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_bands.R
\name{get_posterior_covariance}
\alias{get_posterior_covariance}
\title{#' Posterior inference for covariance surface
#'
#' @param mcmc_results mcmc object
#' @param eigenvals Number of eigenvalues to keep
#' @param zi Covariate vector of interest
#' @param alpha_level Type I error rate
#' @return A cube
#' @export get_posterior_covariance}
\usage{
get_posterior_covariance(mcmc_results, zi, alpha_level = 0.05)
}
\description{
#' Posterior inference for covariance surface
#'
#' @param mcmc_results mcmc object
#' @param eigenvals Number of eigenvalues to keep
#' @param zi Covariate vector of interest
#' @param alpha_level Type I error rate
#' @return A cube
#' @export get_posterior_covariance
}
