% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Posterior_bands.R
\name{get_posterior_subject_bands}
\alias{get_posterior_subject_bands}
\title{Get means and upper/lower credible bands for each subject}
\usage{
get_posterior_subject_bands(
  mcmc_output,
  alpha_level = 0.05,
  mode = "point-wise"
)
}
\arguments{
\item{mcmc_output}{Object returned from run_mcmc()}

\item{alpha_level}{Type I error}

\item{mode}{Type of posterior bands. Default is point-wise. Can specify
simultaneous for approximate simultaneous coverage}
}
\value{
a tibble with columns subject ID, response, mean, lower limit,
upper limit, and time
}
\description{
Posterior processing function for computing means and credible bands. Refer 
to our manuscript for details on credible band calculation
}
