% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Run Markov-Chain Monte-Carlo}
\usage{
run_mcmc(
  response,
  design_mean,
  design_var,
  basis,
  time,
  penalties_mean,
  penalties_var,
  indices_mean,
  indices_var,
  kdim,
  iter,
  burnin,
  thin = 1L,
  var = "unequal"
)
}
\arguments{
\item{response}{N x T response matrix, where N is number of subjects and T is number
of time points. Values can be NA if there's missing data}

\item{design_mean}{N x d_{1} design matrix for mean structure}

\item{design_var}{N x d_{2} design matrix for the covariance structure}

\item{basis}{User generated basis matrix}

\item{time}{vector of time points}

\item{penalties_mean}{List of smoothing penalties for mean structure}

\item{penalties_var}{List of smoothin penalties for covariance structure}

\item{indices_mean}{Maps penalties in the mean structure to beta coefficients}

\item{indices_var}{Maps penalties in the covariance structure to lambda
coefficients}

\item{kdim}{Dimension of latent subspace}

\item{iter}{Number of iterations to run}

\item{burnin}{Number of iterations to use as burn-in. This is only relevant when
passing the returned object into post-processing functions}

\item{thin}{Thinning defaulting to 1}

\item{var}{Can be set to "unequal" to estimate subject-specific measurement
errors or "pooled" to estimate a pooled measurement error variance}
}
\value{
A List containing 3 lists including data, control, and samples.
}
\description{
Generate samples from the posterior distribution. This
algorithm pre-dominantly uses Gibbs sampling
}
