% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Posterior_bands.R
\name{get_posterior_means}
\alias{get_posterior_means}
\title{Get posterior mean function}
\usage{
get_posterior_means(mcmc_output, xi, alpha_level = 0.05, mode = "point-wise")
}
\arguments{
\item{mcmc_output}{object returned from run_mcmc}

\item{xi}{specified covariate value to obtain mean function}

\item{alpha_level}{Type I error}

\item{mode}{Type of posterior bands. Default is point-wise. Can specify
simultaneous for approximate simultaneous coverage}
}
\value{
A tibble with columns for lower bound, mean, upper bound, 
and user specified time grid
}
\description{
Process samples from an object returned from run_mcmc to obtain a posterior
mean with upper/lower bounds. The mean can be evaluated at a specific 
value of covariate.
}
\examples{
# Intercept only model
xi <- c(1)
mean_bands <- get_posterior_means(mcmc_output, xi)
}
