% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_bands.R
\name{get_posterior_subject_bands}
\alias{get_posterior_subject_bands}
\title{Get means and upper/lower credible bands for each subject}
\usage{
get_posterior_subject_bands(mcmc_output, alpha_level = 0.05)
}
\arguments{
\item{mcmc_output}{Object returned from run_mcmc()}

\item{alpha_level}{Type I error}
}
\value{
a tibble with columns subject ID, response, mean, lower limit,
upper limit, and time
}
\description{
Posterior processing function for computing means and credible bands. Refer 
to our manuscript for details on credible band calculation
}
\examples{
subject_bands <- get_posterior_subject_bands(mcmc_output)
subj <- 80
p1 <- subject_bands \%>\% 
  filter(id == subj) \%>\%
 ggplot() +
 geom_point(aes(x = time, y = response)) +
 geom_ribbon(aes(x = time, ymin = lower, ymax = upper), alpha = 0.3) +
 theme_bw()
p1

}
